+++ src/api/server.test.ts
@@ -2,10 +2,10 @@ import { describe, it, expect } from 'vitest';
 import request from 'supertest';
 import { app } from './server.js';
 
-describe('POST /calculate', () => {
+describe('POST /api/v1/calculate', () => {
   it('returns calculated date for FIXED_DAY rule when reference date is before the fixed day', async () => {
     const response = await request(app)
-      .post('/calculate')
+      .post('/api/v1/calculate')
       .send({
         rule: { type: 'FIXED_DAY', dayOfMonth: 20 },
         referenceDate: '2025-01-10',
@@ -21,7 +21,7 @@ describe('POST /calculate', () => {
 
   it('returns calculated date in next month when reference date is after the fixed day', async () => {
     const response = await request(app)
-      .post('/calculate')
+      .post('/api/v1/calculate')
       .send({
         rule: { type: 'FIXED_DAY', dayOfMonth: 5 },
         referenceDate: '2025-01-10',
@@ -36,10 +36,10 @@ describe('POST /calculate', () => {
   });
 });
 
-describe('POST /preview', () => {
+describe('POST /api/v1/preview', () => {
   it('returns next 3 due dates for FIXED_DAY rule (day 20) starting from 2025-01-01', async () => {
     const response = await request(app)
-      .post('/preview')
+      .post('/api/v1/preview')
       .send({
         rule: { type: 'FIXED_DAY', dayOfMonth: 20 },
         from: '2025-01-01',
@@ -60,7 +60,7 @@ describe('POST /preview', () => {
 
   it('returns sequential due dates with clamping for day 31 starting in February', async () => {
     const response = await request(app)
-      .post('/preview')
+      .post('/api/v1/preview')
       .send({
         rule: { type: 'FIXED_DAY', dayOfMonth: 31 },
         from: '2025-02-15',
@@ -87,10 +87,10 @@ describe('POST /preview', () => {
   });
 });
 
-describe('POST /simulate-card', () => {
+describe('POST /api/v1/simulate-card', () => {
   it('returns 3 months of closing/due pairs for range 5-11, offset 8', async () => {
     const response = await request(app)
-      .post('/simulate-card')
+      .post('/api/v1/simulate-card')
       .send({
         closingRangeStart: 5,
         closingRangeEnd: 11,
