+++ src/api/v1/controllers/healthController.ts
@@ -1,0 +1,5 @@
+import type { Request, Response } from 'express';
+
+export function getHealth(_req: Request, res: Response): void {
+  res.json({ status: 'ok' });
+}
